requirements.txt
# python==3.7.3
absl_py==0.8.1
matplotlib==3.1.1
numpy==1.17.3
seaborn==0.9.0
tensorboard==2.1.0
tensorflow-datasets==1.2.0
tensorflow-gpu==2.1.0rc1
typing==3.7.4.1
